Add script to import a project's git submodules into a local gitlab
This script allows a user to import a project's submodules into a local gitlab
instance. This is useful in the hypothetical case where a user is given an
existing repo with pre-initted submodules with an inaccessible remote, but a
copy of the code to be retained in an accessible gitlab instance is desired.

An example invocation of the script would be:

$ ./import.zsh \
	-a 'gitlab-personal-access-token' \
	-r /path/to/project/repo \
	-R gitlab_remote_hostname \
	-u my_username \
	-n name_of_local_repo_value

This invocation would effectively perform the following steps for each submodule
in the repo:

1. Call GitLab API to create the repo at
   my_username/sanitized_version_of_submodule_inner_path
2. git remote add name_of_local_repo_value \
   git@gitlab_remote_hostname:my_username/sanitized_version_of_submodule_inner_path.git
3. git push name_of_local_repo_value

Signed-off-by:	Shawn Webb <shawn.webb@ioactive.com>
Sponsored-by:	IOActive, Inc